reorganized the merged version. Finalized the merge between contents from minbranch and davidbranch
