+++ App.js
@@ -7,13 +7,14 @@ import MapScreen from './src/screens/MapScreen';
 import AccountScreen from './src/screens/AccountScreen';
 import LoginScreen from './src/screens/LoginScreen';
 import RegisterScreen from "./src/screens/RegisterScreen";
-import ForgetPasswordScreen from "./src/screens/ForgetPassword";
+import ForgetPasswordScreen from "./src/screens/ForgetPasswordScreen";
 
 const switchNavigator = createSwitchNavigator({
   loginFLow: createSwitchNavigator(
     {
       Register: RegisterScreen,
       Login: LoginScreen,
+      ForgetPassword: ForgetPasswordScreen
     },
     { initialRouteName: 'Login' }
   ),
@@ -27,19 +28,4 @@ const switchNavigator = createSwitchNavigator({
   })
 });
 
-const navigator = createStackNavigator(
-  {
-    login: LoginScreen,
-    list: ListScreen,
-    register: RegisterScreen,
-    forgot_password: ForgetPasswordScreen
-  },
-  {
-    initialRouteName: 'login',
-    defaultNavigationOptions: {
-      title: 'StreetFood'
-    }
-  }
-);
-
-export default createAppContainer(navigator);
+export default createAppContainer(switchNavigator);
